Simplified entire system, fixed namespaces
Removed the seperate files for AccessMovieData/AccessPersonData and combined them, added namespaces to reachability. Removed unused Person.json file from testing
