--- Person.json
@@ -1,23 +1,0 @@
-[
-  {
-    "Email": "chung@gamil",
-    "Name": "chung",
-    "Password": "pass",
-    "IsAdmin": false,
-    "Reservations": []
-  },
-  {
-    "Email": "chung2",
-    "Name": "chung2",
-    "Password": "pas",
-    "IsAdmin": false,
-    "Reservations": []
-  },
-  {
-    "Email": "maria",
-    "Name": "maria",
-    "Password": "pas",
-    "IsAdmin": false,
-    "Reservations": []
-  }
-]
